refactor(album) : MP-91 Add some condition to complete test case

+++ api/album/edit.php
@@ -20,21 +20,36 @@ if ($_SERVER['REQUEST_METHOD'] == 'POST') {
 
             if (isset($_POST['albumName'])) {
                 $albumName = $_POST['albumName'];
-                $query = "UPDATE album SET albumName = '$albumName' WHERE albumId = '$albumId'";
-                if (mysqli_query($database->connection, $query)) $checkName = true;
+                if (!empty($albumName)) {
+                    $query = "UPDATE album SET albumName = '$albumName' WHERE albumId = '$albumId'";
+                    if (mysqli_query($database->connection, $query)) $checkName = true;
+                }
             }
 
             if (isset($_POST['accountIds'])) {
                 $accountIds = $_POST['accountIds'];
-                $query = "DELETE FROM album_artist WHERE albumId = '$albumId'";
-                $execute = mysqli_query($database->connection, $query);
-                foreach ($accountIds as $index => $accountId) {
-                    $query = "INSERT INTO album_artist (albumId, accountId) VALUES ('$albumId', '$accountId')";
-                    if (mysqli_query($database->connection, $query)) $checkArtist = true;
+                if (!in_array("", $accountIds)) {
+                    $accountExist = false;
+                    foreach ($accountIds as $index => $accountId) {
+                        $queryCheck = "SELECT * FROM account WHERE accountId = '$accountId'";
+                        $execute = mysqli_query($database->connection, $queryCheck);
+                        $accountExist = mysqli_num_rows($execute) > 0 ? true : false;
+                        if (!$accountExist) {
+                            break;
+                        }
+                    }
+                    if ($accountExist) {
+                        $query = "DELETE FROM album_artist WHERE albumId = '$albumId'";
+                        $execute = mysqli_query($database->connection, $query);
+                        foreach ($accountIds as $index => $accountId) {
+                            $query = "INSERT INTO album_artist (albumId, accountId) VALUES ('$albumId', '$accountId')";
+                            if (mysqli_query($database->connection, $query)) $checkArtist = true;
+                        }
+                    }
                 }
             }
 
-            if (isset($_FILES['albumPhoto'])) {
+            if (isset($_FILES['albumPhoto']["tmp_name"])) {
                 $querySelect = "SELECT * FROM album WHERE albumId = '$albumId'";
                 $executeSelect = mysqli_query($database->connection, $querySelect);
                 $check = mysqli_affected_rows($database->connection);
